Move copyMultipleImages to copy/multiple.go
Only moves unchanged code, should not change behavior.
